pyproject.toml
[project]
name = "illustrator"
version = "0.1.0"
description = "A LangGraph application for analyzing manuscripts and generating illustrations using AI"
authors = [
    { name = "David James Lennon", email = "david@othertales.co" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    # LangGraph and LangChain ecosystem
    "langgraph>=0.1.0",
    "langgraph-sdk>=0.1.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "python-dotenv>=1.0.1",
    # Persistence and data access
    "sqlalchemy>=2.0.30",
    # NLP and text processing
    "spacy>=3.7.0",
    "textblob>=0.17.1",
    "nltk>=3.8.1",
    "transformers>=4.40.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    # Image generation APIs
    "openai>=1.47.0",
    "google-cloud-aiplatform>=1.60.0",
    "huggingface_hub>=0.24.0",
    "langchain-huggingface>=0.1.0",
    "replicate>=0.24.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    # CLI and utilities
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pillow>=10.4.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "httpx>=0.27.0",
    # Exports
    "python-docx>=1.1.0",
    "reportlab>=4.2.0",
    "weasyprint>=61.0",
]
db = [
    "psycopg2-binary>=2.9.9"
]
dev = [
    "beautifulsoup4>=4.12.0",
    "langsmith>=0.1.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
    "selenium>=4.21.0",
]

[project.scripts]
illustrator = "illustrator.cli:main"

[build-system]
requires = ["setuptools>=80.9.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["illustrator"]
[tool.setuptools.package-dir]
"illustrator" = "src/illustrator"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
include = ["*.py", "*.pyi", "*.ipynb"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_errors = true
